projects/03-walmart-sales-forecasting-saas/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic-settings>=2.1.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.2
python-dotenv>=1.0.0
# ML Forecasting Models (optional - install if needed)
prophet>=1.1.5
statsmodels>=0.14.1
# pmdarima>=2.0.4
# Data Adapter Dependencies
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
python-dateutil>=2.8.2
pandera>=0.18.0
